+++ Modules/DataModel/Sources/PocketCastsDataModel/Public/Model/Episode+fromDatabase.swift
@@ -42,6 +42,7 @@ extension Episode {
         episode.excludeFromEpisodeLimit = rs.bool(forColumn: "excludeFromEpisodeLimit")
         episode.starredModified = rs.longLongInt(forColumn: "starredModified")
         episode.deselectedChapters = rs.string(forColumn: "deselectedChapters")
+        episode.deselectedChaptersModified = rs.longLongInt(forColumn: "deselectedChaptersModified")
         return episode
     }
 }