+++ Modules/Server/Sources/PocketCastsServer/Public/Sync/SyncTask+LocalChanges.swift
@@ -72,6 +72,10 @@ extension SyncTask {
                 episodeRecord.isDeleted.value = episode.archived
                 episodeRecord.isDeletedModified.value = episode.archivedModified
             }
+            if let deselectedChapters = episode.deselectedChapters {
+                episodeRecord.deselectedChapters = deselectedChapters
+                episodeRecord.deselectedChaptersModified.value = episode.deselectedChaptersModified
+            }
 
             var apiRecord = Api_Record()
             apiRecord.episode = episodeRecord
@@ -231,6 +235,7 @@ private extension Podcast {
         settings.addToUpNextPosition.update(self.settings.$addToUpNextPosition)
         settings.episodesSortOrder.update(self.settings.$episodesSortOrder)
         settings.episodeGrouping.update(self.settings.$episodeGrouping)
+        settings.showArchived.update(self.settings.$showArchived)
         settings.autoArchive.update(self.settings.$autoArchive)
         settings.autoArchivePlayed.update(self.settings.$autoArchivePlayed)
         settings.autoArchiveInactive.update(self.settings.$autoArchiveInactive)