+++ Pocket Casts Watch App/WatchHostingController.swift
@@ -23,7 +23,7 @@ class WatchHostingController: WKHostingController<AnyView>, Restorable {
 
     func restoreName() -> String? {
         switch controllerType {
-        case .downloads, .podcasts, .files, .filter, .upnext, .nowPlaying:
+        case .downloads, .podcasts, .files, .filter, .upnext, .nowPlaying, .filterList:
             return Self.interfaceName
         default:
             return nil