+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -----
 - Profile bookmark screen where all user bookmarks can be managed. [#1500]
 - Fix the player getting stuck and not reopening [#1529]
+- Visually mark played episodes in search results [#928]
 
 7.59
 -----